support fractional share quantities in holdings and guidance

+++ backend/app/models/position.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Float, ForeignKey, Integer
+from sqlalchemy import String, Float, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db.base import Base, UUIDMixin, TimestampMixin
 
@@ -7,7 +7,7 @@ class Position(Base, UUIDMixin, TimestampMixin):
     __tablename__ = "positions"
 
     symbol: Mapped[str] = mapped_column(String(10), nullable=False, index=True, unique=True)
-    quantity: Mapped[int] = mapped_column(Integer, nullable=False)
+    quantity: Mapped[float] = mapped_column(Float, nullable=False)
     avg_cost: Mapped[float] = mapped_column(Float, nullable=False)
     current_price: Mapped[float | None] = mapped_column(Float, nullable=True)
     stop_level: Mapped[float | None] = mapped_column(Float, nullable=True)
