+++ backend/app/services/broker/base.py
@@ -20,7 +20,7 @@ class AccountSummary:
 @dataclass
 class PositionDto:
     symbol: str
-    quantity: int
+    quantity: float
     avg_cost: float
     current_price: float | None
     unrealized_pnl: float | None
